refactor(1013): remove SFPTransceiverDataProcessor facade (Cat B, position 27)
Extract SFPTransceiverDataProcessor from MistHelper.py into
src/reports/sfp_transceiver_data_processor.py using the SC-001 facade
pattern. Lazy MistHelper import via importlib avoids circular init.
Callers continue to reach the class through the re-export alias in
MistHelper.py.

- src/reports/sfp_transceiver_data_processor.py (new)
- MistHelper.py: re-export alias + comment breadcrumb
- pyproject.toml: coverage-omit entry
