+++ pyproject.toml
@@ -243,6 +243,7 @@ omit = [
     "src/export/site_config_exporter.py",
     "src/firmware/firmware_manager.py",
     "src/gateway/gateway_ha_exporter.py",
+    "src/reports/sfp_transceiver_data_processor.py",
     "src/reports/wired_client_manufacturer_report_generator.py",
     "src/site/bulk_radius_wlan_config_manager.py",
     "src/ui/tui.py",
